Initial work on getting `moved` supported in Feature.yaml
This can be used to rename keys, as long as the new name is not already
configured:

```
moved:
  - from: old.key.name
    to: new.name
```

It should allow viewing and managing the new keys for a rollout, without
affecting other features until the rollout is complete.
While it's a rollout the change is virtual (just renamed in the
backend). When the rollout is complete, all keys will be renamed
permanently in the database.