Fix environment resolver and cluster upgrader
This commit fixes an issue in the environment resolver where it was not properly handling nil progress. It now checks if the progress is nil before iterating over the metrics.

In the cluster upgrader, the code has been added to upgrade node pools to the specified version. However, the code is currently commented out and needs to be uncommented to perform the actual upgrade.